refactor: replace UI barrel imports with direct imports
