+++ components/layout/nav.tsx
@@ -5,7 +5,7 @@ import { useScopedI18n } from "@/lib/locales/client";
 import { cn } from "@/lib/utils";
 import Link from "next/link";
 import { useParams, usePathname } from "next/navigation";
-import { buttonVariants } from "../ui";
+import { buttonVariants } from "../ui/button";
 
 export default function Nav() {
   const t = useScopedI18n("Nav");
