--- components/ui/index.ts
@@ -1,24 +1,0 @@
-import { BadgeLink } from "./badge";
-import { Button, type ButtonProps, buttonVariants } from "./button";
-
-import { CopyButton } from "./copy-button";
-import { Icons } from "./icons";
-import { ScrollArea, ScrollBar } from "./scroll-area";
-import { Section } from "./section";
-import { Snippet } from "./snippet";
-import { Spacing } from "./spacing";
-
-export {
-  Button,
-  buttonVariants,
-  CopyButton,
-  Icons,
-  Section,
-  Spacing,
-  Snippet,
-  BadgeLink,
-  ScrollArea,
-  ScrollBar,
-};
-
-export type { ButtonProps };
